% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consolidate_partial_results.R
\name{consolidate.partial.results}
\alias{consolidate.partial.results}
\title{Consolidate results from partial files}
\usage{
consolidate.partial.results(folder = "./nreps_files")
}
\arguments{
\item{folder}{folder where the partial files are located.}
}
\value{
a list object containing the following fields:
\itemize{
\item \code{data.raw} - data frame containing all observations generated
\item \code{data.summary} - data frame summarizing the experiment.
\item \code{N} - number of instances sampled
\item \code{total.runs} - total number of algorithm runs performed
\item \code{instances.sampled} - names of the instances sampled
}
}
\description{
Consolidates results from a set of partial files (each generated by an
individual call to \code{\link[=calc_nreps]{calc_nreps()}}) into a single output structure, similar
(but not identical) to the output of \code{\link[=run_experiment]{run_experiment()}}.
}
