% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummyalgo.R
\name{dummyalgo}
\alias{dummyalgo}
\title{Dummy algorithm routine to test the sampling procedures}
\usage{
dummyalgo(
  distribution.fun = "rnorm",
  distribution.pars = list(mean = 0, sd = 1),
  instance = list(FUN = "dummyinstance", distr = "rnorm", mean = 0, sd = 1)
)
}
\arguments{
\item{distribution.fun}{name of a function that generates random values
according to a given distribution, e.g., "rnorm", "runif", "rexp" etc.}

\item{distribution.pars}{list of named parameters required by the function
in \code{distribution.fun}. Parameter \code{n} (number of points to
generate) is unnecessary (this routine always forces \code{n = 1}).}

\item{instance}{instance parameters (see \code{Details}).}
}
\value{
a list object with a single field \code{$value}, containing a scalar
numerical value distributed as described at the end of \code{Details}.
}
\description{
This is a dummy algorithm routine to test the sampling procedures, in
combination with \code{\link[=dummyinstance]{dummyinstance()}}.
\code{dummyalgo()} receives two parameters that determine the distribution of
performances it will exhibit on a hypothetical problem class:
\code{distribution.fun} (with the name of a random number generation function,
e.g. \code{rnorm}, \code{runif}, \code{rexp} etc.); and \code{distribution.pars}, a named list of
parameters to be passed on to \code{distribution.fun}.
The third parameter is an instance object (see \code{\link[=calc_nreps]{calc_nreps()}} for details),
which is a named list with the following fields:
\itemize{
\item\code{FUN = "dummyinstance"} - must always be "dummyinstance" (will
be ignored otherwise).
\item\code{distr} - the name of a random number generation function.
\item\code{...} - other named fields with parameters to be passed down
to the function in \code{distr}.
}
}
\details{
\code{distribution.fun} and \code{distribution.pars} regulate the mean performance of
the dummy algorithm on a given (hypothetical) problem class, and the
between-instances variance of performance. The instance specification in
\code{instance} regulates the within-instance variability of results. Ideally the
distribution parameters passed to the \code{instance} should result in a
within-instance distribution of values with zero mean, so that the mean of
the values returned by \code{dummyalgo} is regulated only by \code{distribution.fun}
and \code{distribution.pars}.

The value returned by dummyalgo is sampled as follows:
\preformatted{
   offset <- do.call(distribution.fun, args = distribution.pars)
   y <- offset + do.call("dummyinstance", args = instance)
}
}
\examples{
# Make a dummy instance with a centered (zero-mean) exponential distribution:
instance = list(FUN = "dummyinstance", distr = "rexp", rate = 5, bias = -1/5)

# Simulate a dummy algorithm that has a uniform distribution of expected
# performance values, between -25 and 50.
dummyalgo(distribution.fun = "runif",
          distribution.pars = list(min = -25, max = 50),
          instance = instance)

}
\seealso{
\code{\link[=dummyinstance]{dummyinstance()}}
}
\author{
Felipe Campelo (\email{fcampelo@ufmg.br},
\email{f.campelo@aston.ac.uk})
}
