% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_perc.R
\name{est_perc}
\alias{est_perc}
\title{Percent differences estimator}
\usage{
est_perc(x1, x2)
}
\arguments{
\item{x1}{vector of observations}

\item{x2}{vector of observations}
}
\value{
Estimated value of the statistic
}
\description{
Calculates the sample estimator of \eqn{(mu2 - mu1) / mu1}, where
\eqn{mu1, mu2} are the means of the populations that generated the sample
vectors \eqn{x1, x2}.
}
\section{References}{

\itemize{
\item F. Campelo, F. Takahashi:
Sample size estimation for power and accuracy in the experimental
comparison of algorithms (submitted, 2017).
}
}

\author{
Felipe Campelo (\email{fcampelo@ufmg.br})
}
