% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{my.SANN}
\alias{my.SANN}
\title{Simulated annealing (for testing/examples)
Adapted from stats::optim}
\usage{
my.SANN(Temp, budget, instance)
}
\arguments{
\item{Temp}{controls the "SANN" method. It is the starting temperature for
the cooling schedule.}

\item{budget}{stop criterion: total number of function evaluations}

\item{instance}{instance object (see \code{\link[=calc_nreps2]{calc_nreps2()}} for details)}
}
\description{
Simulated annealing (for testing/examples)
Adapted from stats::optim
}
