% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_caiser_powercurve.R
\name{plot.caiser.powercurve}
\alias{plot.caiser.powercurve}
\title{plot.caiser.powercurve}
\usage{
\method{plot}{caiser.powercurve}(object, ...)
}
\arguments{
\item{object}{list object of class \emph{caiser.powercurve}
(generated by \code{\link[=calc_power_curve]{calc_power_curve()}})}

\item{...}{other parameters to be passed down to specific plotting
functions (currently unused)}
}
\description{
S3 method for plotting \emph{caiser.powercurve} objects (the output of
\code{\link[=calc_power_curve]{calc_power_curve()}}).
}
\examples{
my.cpc <- calc_power_curve(ninstances = 10)
plot(my.cpc)

}
