% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_caiser.R
\name{plot.CAISEr}
\alias{plot.CAISEr}
\title{plot.CAISEr}
\usage{
\method{plot}{CAISEr}(x, y = NULL, ..., latex = FALSE,
  reorder = FALSE, show.text = TRUE, layout = NULL)
}
\arguments{
\item{x}{list object of class \emph{CAISEr}.}

\item{y}{unused. Included for consistency with generic \code{plot} method.}

\item{...}{other parameters to be passed down to specific
plotting functions (currently unused)}

\item{latex}{logical: should labels be formatted for LaTeX? (useful for
later saving using library \code{TikzDevice})}

\item{reorder}{logical: should the comparisons be reordered alphabetically?}

\item{show.text}{logical: should text be plotted?}

\item{layout}{optional parameter to override the layout of the plots (see
\code{gridExtra::arrangeGrobs()} for details.}
}
\value{
list of \code{ggplot} objects (invisibly), with the three individual plots
generated (useful e.g., for extending limits or refining the plots)
}
\description{
S3 method for plotting \emph{CAISEr} objects output by \code{\link[=run_experiment]{run_experiment()}}).
}
