% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_caiser_powercurve.R
\name{summary.caiser.powercurve}
\alias{summary.caiser.powercurve}
\title{summary.caiser.powercurve}
\usage{
\method{summary}{caiser.powercurve}(object, ...)
}
\arguments{
\item{object}{list object of class \emph{caiser.powercurve}
(generated by \code{\link[=calc_power_curve]{calc_power_curve()}})}

\item{...}{other parameters to be passed down to specific summary
functions (currently unused)}
}
\description{
S3 method for summarizing \emph{caiser.powercurve} objects (the output of
\code{\link[=calc_power_curve]{calc_power_curve()}}).
}
\examples{
my.cpc <- calc_power_curve(ninstances = 10)
summary(my.cpc)

}
