% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_CAISErPowercurve.R
\name{summary.CAISErPowercurve}
\alias{summary.CAISErPowercurve}
\title{summary.CAISErPowercurve}
\usage{
\method{summary}{CAISErPowercurve}(object, highlights = c(0.5, 0.8, 0.95),
  ...)
}
\arguments{
\item{object}{list object of class \emph{caiser.powercurve}
(generated by \code{\link[=calc_power_curve]{calc_power_curve()}})}

\item{highlights}{specific power values to highlight}

\item{...}{other parameters to be passed down to specific summary
functions (currently unused)}
}
\description{
S3 method for summarizing \emph{caiser.powercurve} objects (the output of
\code{\link[=calc_power_curve]{calc_power_curve()}}).
}
\examples{
my.cpc <- calc_power_curve(ninstances = 10)
summary(my.cpc)

}
