% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_nreps.R
\name{summary.nreps}
\alias{summary.nreps}
\title{summary.nreps}
\usage{
\method{summary}{nreps}(object, ...)
}
\arguments{
\item{object}{list object of class \emph{nreps}
(generated by \code{\link[=calc_nreps]{calc_nreps()}})}

\item{...}{other parameters to be passed down to specific
summary functions (currently unused)}
}
\description{
S3 method for summarizing \emph{nreps} objects output by \code{\link[=calc_nreps]{calc_nreps()}}).
}
