% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{TSP.dist}
\alias{TSP.dist}
\title{TSP instance generator (for testing/examples)
Adapted from stats::optim}
\usage{
TSP.dist(x, mydist)
}
\arguments{
\item{x}{a valid closed route for the TSP instance}

\item{mydist}{object of class \emph{dist} defining the TSP instance}
}
\description{
TSP instance generator (for testing/examples)
Adapted from stats::optim
}
