% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dummyalgo}
\alias{dummyalgo}
\title{Dummy algorithm routine to test the sampling procedures}
\usage{
dummyalgo(distribution.fun = "rnorm", distribution.pars = list(mean = 0, sd
  = 1), instance)
}
\arguments{
\item{distribution.fun}{name of a function that generates random values
according to a given distribution, e.g., "rnorm", "runif", "rexp" etc.}

\item{distribution.pars}{list of named parameters required by the function
in \code{distribution.fun}. Parameter \code{n} (number of points to
generate) is unnecessary (this routine always considers \code{n = 1}).}

\item{instance}{instance parameters. This parameter is always ignored, and
was only included for compatibility with \code{run_nreps2} and other \code{nreps}
functions.}
}
\value{
a list object with a single field \code{$value}, containing a value
distributed according to \code{distribution.fun} and \code{distribution.pars}.
}
\description{
This is a dummy algorithm routine to test the sampling procedures. It
basically returns values according to the distribution function given
in \code{distribution.fun} with parameters given by \code{distribution.pars}.
}
\author{
Felipe Campelo (\email{fcampelo@ufmg.br})
}
