% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dummyinstance}
\alias{dummyinstance}
\title{Dummy instance (for testing only) - a function that does nothing and returns
nothing}
\usage{
dummyinstance()
}
\description{
Dummy instance (for testing only) - a function that does nothing and returns
nothing
}
