% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSP_dist.R
\name{TSP.dist}
\alias{TSP.dist}
\title{TSP instance generator (for testing/examples)}
\usage{
TSP.dist(x, mydist)
}
\arguments{
\item{x}{a valid closed route for the TSP instance}

\item{mydist}{object of class \emph{dist} defining the TSP instance}
}
\description{
Adapted from stats::optim(). Check their documentation / examples for
details.
}
